updates: fixed everything :)
