+++ CSI-340_ChenRoblesWu_Website/CSI-340_ChenRoblesWu_Website/Properties/Resources.resx
@@ -118,6 +118,6 @@
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
   <data name="ConnectionString" xml:space="preserve">
-    <value>Data Source ="(LocalDb)\MSSQLLocalDb"; database = "gravity_books";integrated security=SSPI;</value>
+    <value>Data Source ="DESKTOP-4MHOVL2"; database = "gravity_books";integrated security=SSPI;</value>
   </data>
 </root>
