--- CSI-340_ChenRoblesWu_Website/CSI-340_ChenRoblesWu_Website/Views/Home/Browse.cshtml
@@ -1,6 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Browse";
-}
-<h1>@ViewData["Title"]</h1>
-
-<p>Browse our book inventory here</p>
