+++ CSI-340_ChenRoblesWu_Website/CSI-340_ChenRoblesWu_Website/Views/Home/Index.cshtml
@@ -1,73 +1,6 @@
 ﻿@{
-    ViewData["Title"] = "Displaying Address Data";
-
-    string[]TableHeaders = new string[]
-    {
-
-       "BookId"
-      ,"title"
-      ,"isbn13"
-      ,"LanguageId"
-      ,"NumPages"
-      ,"PublicationDate"
-      ,"PublisherId"
- 
-    };
-
+    ViewData["Title"] = "Index";
 }
-<div class ="table">
-    <table class ="table table-border table-hover">
-        <thead>
-            <tr>
-                @{
-                    foreach(var head in TableHeaders)
-                    {
-                        <th>
-                            @head
-                        </th>
-                        
-                    }
-                }
-            </tr>
 
 
-        </thead>
-
-        @{
-            if(Model != null)
-            {
-                foreach (var Data in Model)
-                {
-                    <tr>
-                        
-                        <td>
-                            @Data.BookId;
-                        </td>
-                         <td>
-                            @Data.Title;
-                        </td>
-                         <td>
-                            @Data.isbn13;
-                        </td>
-                         <td>
-                            @Data.LanguageId;
-                        </td>
-                        <td>
-                            @Data.NumPages;
-                        </td>                      
-                         <td>
-                            @Data.PublicationDate;
-                        </td>
-                         <td>
-                            @Data.PublisherId;
-                        </td>
-
-                    </tr>
-         
-                 }
-            }
-        }
-
-    </table>    
-
-</div>
+<p>Welcome to our bookstore!</p>
